+++ R/MsBackendMetaboLights.R
@@ -100,6 +100,8 @@
 #'
 #' @importClassesFrom Spectra MsBackendDataFrame
 #'
+#' @importFrom S4Vectors DataFrame
+#'
 #' @exportClass MsBackendMetaboLights
 #'
 #' @author Philippine Louail, Johannes Rainer
@@ -134,6 +136,10 @@ MsBackendMetaboLights <- function() {
 #' @rdname MsBackendMetaboLights
 #'
 #' @importMethodsFrom ProtGenerics backendInitialize
+#'
+#' @importMethodsFrom Spectra backendInitialize
+#'
+#' @importMethodsFrom Spectra [
 #'
 #' @importMethodsFrom ProtGenerics dataOrigin
 #'
@@ -206,21 +212,40 @@ setMethod(
 #' The functions listed here allow to query and retrieve information of a
 #' data set/experiment from MetaboLights.
 #'
-#' - `mtbls_ftp_path`: returns the FTP path for a provided MetaboLights ID.
+#' - `mtbls_ftp_path()`: returns the FTP path for a provided MetaboLights ID.
 #'   With `mustWork = TRUE` (the default) the function throws an error if
 #'   the path is not accessible (either because the data set does not exist or
 #'   no internet connection is available). The function returns a
 #'   `character(1)` with the FTP path to the data set folder.
 #'
-#' - `mtbls_list_files`: returns the available files (and directories) for the
+#' - `mtbls_cached_data_files()`: lists locally cached data files from
+#'   MetaboLights. Since this function evaluates only local content it does not
+#'   require an internet connection. With the default parameters all available
+#'   data files are listed. The parameters can be used to restrict the lookup.
+#'
+#' - `mtbls_list_files()`: returns the available files (and directories) for the
 #'   specified MetaboLights data set (i.e., the FTP directory content of the
 #'   data set). The function returns a `character` vector with the relative
 #'   file names to the absolute FTP path (`mtbls_ftp_path()`) of the data set.
 #'   Parameter `pattern` allows to filter the file names and define which
 #'   file names should be returned.
 #'
+#' - `mtbls_sync_data_files()`: synchronize data files of a specifies
+#'   MetaboLights data set eventually downloading and locally caching them.
+#'   Parameter `fileName` allows to specify names of selected data files to
+#'   sync.
+#'
 #' @param x `character(1)` with the ID of the MetaboLights data set (usually
 #'     starting with a *MTBLS* followed by a number).
+#'
+#' @param mtblsId `character(1)` with the ID of a single MetaboLights data
+#'     set/experiment.
+#'
+#' @param assayName `character` with the file names of assay files of the data
+#'     set. If not provided (`assayName = character()`, the default), MS data
+#'     files of all data set's assays are loaded. Use
+#'     `mtbls_list_files(<MetaboLights ID>, pattern = "^a_")` to list all
+#'     available assay files of a data set `<MetaboLights ID>`.
 #'
 #' @param mustWork for `mtbls_ftp_path()`: `logical(1)` whether the validity of
 #'     the path should be verified or not. By default (with `mustWork = TRUE`)
@@ -228,11 +253,17 @@ setMethod(
 #'     if the folder can not be accessed (e.g. if no internet connection is
 #'     available).
 #'
-#' @param pattern for `mtbls_list_files()`: `character(1)` defining a pattern
+#' @param pattern for `mtbls_list_files()`, `mtbls_sync_data_files()` and
+#'     `mtbls_cached_data_files()`: `character(1)` defining a pattern
 #'     to filter the file names, such as `pattern = "^a_"` to retrieve the
 #'     file names of all assay files of the data set (i.e., files with a name
 #'     starting with `"a_"`). This parameter is passed to the [grepl()]
 #'     function.
+#'
+#' @param fileName for `mtbls_sync_data_files()` and
+#'     `mtbls_cached_data_files()`: optional `character`
+#'     defining the names of specific data files of a data set that should be
+#'     downloaded and cached.
 #'
 #' @return
 #'
@@ -240,6 +271,9 @@ setMethod(
 #'   data set on the MetaboLights ftp server.
 #' - For `mtbls_list_files()`: `character` with the names of the files in the
 #'   data set's base ftp directory.
+#' - For `mtbls_sync_data_files()` and `mtbls_cached_data_files()`: a
+#'   `data.frame` with the MetaboLights ID, the assay name(s) and remote and
+#'   local file names of the synchronized data files.
 #'
 #' @author Johannes Rainer, Philippine Louail
 #'
@@ -259,6 +293,9 @@ setMethod(
 #' a <- read.table(paste0(mtbls_ftp_path("MTBLS2"), afiles[1L]),
 #'     header = TRUE, sep = "\t", check.names = FALSE)
 #' head(a)
+#'
+#' ## List all available files
+#' mtbls_cached_data_files()
 NULL
 
 #' @rdname MetaboLights-utils
@@ -352,6 +389,31 @@ mtbls_list_files <- function(x = character(), pattern = NULL) {
 ##
 ################################################################################
 
+#' @rdname MetaboLights-utils
+#'
+#' @export
+mtbls_sync_data_files <- function(mtblsId = character(),
+                                  assayName = character(),
+                                  pattern = "mzML$|CDF$|cdf$|mzXML$",
+                                  fileName = character()) {
+    if (!length(mtblsId))
+        stop("No MetaboLights data set ID provided with parameter 'mtblsId'")
+    invisible(.mtbls_data_files(mtblsId, assayName, pattern, fileName))
+}
+
+#' @rdname MetaboLights-utils
+#'
+#' @export
+mtbls_cached_data_files <- function(mtblsId = character(),
+                                    assayName = character(),
+                                    pattern = "*", fileName = character()) {
+    res <- .mtbls_data_files_offline(mtblsId = mtblsId, assayName = assayName,
+                                     pattern = pattern)
+    if (length(fileName))
+        res <- res[basename(res$derived_spectral_data_file) %in% fileName, ]
+    else res
+}
+
 #' Get information on data files for a given MTBLS ID/assay eventually
 #' downloading and caching them. This function needs an active internet
 #' connection as it queries the MTBLS ftp server for available data files
@@ -379,7 +441,8 @@ mtbls_list_files <- function(x = character(), pattern = NULL) {
 #'
 #' @noRd
 .mtbls_data_files <- function(mtblsId = character(), assayName = character(),
-                              pattern = "mzML$|CDF$|mzXML$") {
+                              pattern = "mzML$|CDF$|mzXML$",
+                              fileName = character()) {
     assays <- .mtbls_assay_list(mtblsId)
     anames <- names(assays)
     if (length(assayName)) {
@@ -404,6 +467,13 @@ mtbls_list_files <- function(x = character(), pattern = NULL) {
             message("Used data files from the assay's column \"Raw Spectral ",
                     "Data File\" since none were available in column ",
                     "\"Derived Spectral Data File\".")
+    }
+    if (length(fileName)) {
+        keep <- basename(ffiles) %in% fileName
+        if (!any(keep))
+            stop("None of the 'fileName' found in data set \"", mtblsId, "\"")
+        ffiles <- ffiles[keep]
+        dfiles <- lapply(dfiles, function(z) z[basename(z) %in% fileName])
     }
     ## Cache files
     bfc <- BiocFileCache()